requirements.txt
python-doctr[torch]
streamlit==1.39.0
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
reportlab==4.2.5
pillow==10.3.0
scikit-learn==1.5.2
opencv-python-headless==4.10.0.84
pytesseract==0.3.13